requirements.txt
Pillow==7.2.0
numpy==1.18.5
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1